+++ diary-server/src/main/java/space/maxus/dnevnik/controllers/request/create/MarkCreateRequest.java
@@ -9,8 +9,8 @@ import java.util.UUID;
 public class MarkCreateRequest {
     private String message;
     private Mark.Kind kind = Mark.Kind.GENERIC;
-    private byte value;
-    private byte index = 1;
+    private int value;
+    private int index = 1;
     private long lessonId;
     private UUID studentId;
 }
